add ControlBtnRegistrationFormModal
