+++ src/components/shared/Header/Header.jsx
@@ -1,10 +1,10 @@
 import { useTranslations } from "next-intl";
 import Logo from "../Logo/Logo";
-import MainButton from "../MainButton/MainButton";
 import LangDropdown from "../LangDropdown/LangDropdown";
 import Menu from "./Menu/Menu";
 import styles from "./Header.module.scss";
 import Burger from "./Burger/Burger";
+import ControlBtnModalPayment from "../controlButtons/ControlBtnModalPayment";
 
 const Header = () => {
   const t = useTranslations("Header");
@@ -14,9 +14,9 @@ const Header = () => {
         <Logo variant="header" className={styles.logo} />
         <Menu />
         <div className={styles.actions}>
-          <MainButton className={styles.headerBtn}>
+          <ControlBtnModalPayment className={styles.headerBtn}>
             {t("btn_support_project")}
-          </MainButton >
+          </ControlBtnModalPayment>
           <LangDropdown />
           <Burger />
         </div>
