+++ src/components/shared/controlButtons/ControlBtnRegistrationFormModal.jsx
@@ -1,0 +1,12 @@
+"use client";
+import useStateModal from "@/src/state/stateRegistrationFormModal";
+import MainButton from "../MainButton/MainButton";
+
+export default function ControlBtnRegistrationFormModal({children, className}) {
+  const open = useStateModal((state) => state.open)
+  return <MainButton 
+    className={className} 
+    onClick={open}>
+      {children}
+  </MainButton>
+}
