+++ src/state/stateRegistrationFormModal.js
@@ -1,0 +1,9 @@
+import { create } from 'zustand';
+
+const stateRegistrationFormModal = create((set) => ({
+  isOpen: false,
+  open: () => set({ isOpen: true }),
+  close: () => set({ isOpen: false }),
+}))
+
+export default stateRegistrationFormModal
