--- src/state/useStateModal.js
@@ -1,9 +1,0 @@
-import { create } from 'zustand'
-
-const useStateModal = create((set) => ({
-  isOpen: false,
-  open: () => set({ isOpen: true }),
-  close: () => set({ isOpen: false }),
-}))
-
-export default useStateModal
